api/requirements.txt
flask
flask-cors
kafka-python
mysql-connector-python
